+++ tools/reorder-localizations.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+
+set -eo pipefail
+
+cd ./reorder-localizations
+./reorder.sh
+node ./format-strings.mjs